solucionado la funcion de costo

+++ agente/src/frsf/cidisi/exercise/diagrama/search/actions/Apagar.java
@@ -47,7 +47,7 @@ public class Apagar extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(1);
+        return new Double(0);
     }
 
     /**
