+++ agente/src/frsf/cidisi/exercise/diagrama/search/actions/SubirOeste.java
@@ -68,7 +68,7 @@ public class SubirOeste extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(1);
+        return new Double(0);
     }
 
     /**
